+++ api/database/migrations/2025_10_07_213921_create_user_balances_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('user_balances', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->integer('amount'); // Valor em centavos
+            $table->decimal('amount', 13, 2); // Valor em reais
             $table->string('description');
             $table->timestamps();
         });
